Deploy archive

+++ 2-do_deploy_web_static.py
@@ -1,46 +1,40 @@
 #!/usr/bin/python3
-# makes tgz archive
-import tarfile
-from fabric.api import local, env, run, put
+""" Function that compress a folder """
 from datetime import datetime
+from fabric.api import *
+import shlex
 import os
 
 
-env.hosts = ['54.197.73.242', '54.208.60.185']
-env.user = 'ubuntu'
-env.key_filename = '~/.ssh/rsa'
-
-
-def do_pack():
-    # packs file
-    try:
-        name = "web_static_" + datetime.now().strftime("%Y%m%d%H%M%S")
-        local('mkdir -p versions')
-        local("tar -cvzf versions/{}.tgz {}".format(
-            name, "web_static/"))
-        size = os.path.getsize("./versions/{}.tgz".format(name))
-        print("web_static packed: versions/{}.tgz -> {}Bytes".format(
-            name, size))
-        return "versions/{}.tgz".format(name)
-    except:
-        return None
+env.hosts = ['	54.197.73.242' '54.208.60.185']
+env.user = "ubuntu"
 
 
 def do_deploy(archive_path):
-    fd = archive_path.split('/')[1]
+    """ Deploys """
+    if not os.path.exists(archive_path):
+        return False
     try:
-        put(archive_path, '/tmp/{}'.format(fd))
-        run('mkdir -p /data/web_static/releases/{}'.format(fd))
-        run('tar -xzf /tmp/{} -C /data/web_static/releases/{}'.format(fd, fd))
-        run('rm /tmp/{}'.format(fd))
-        run('mv /data/web_static/releases/{}/web_static/*\
-        /data/web_static/releases/{}/'.format(fd, fd))
-        run('rm -rf /data/web_static/releases/{}/web_static'.format(fd))
-        run('rm -rf /data/web_static/current')
-        run('ln -s /data/web_static/releases/{}/\
-        /data/web_static/current'.format(fd))
+        name = archive_path.replace('/', ' ')
+        name = shlex.split(name)
+        name = name[-1]
+
+        wname = name.replace('.', ' ')
+        wname = shlex.split(wname)
+        wname = wname[0]
+
+        releases_path = "/data/web_static/releases/{}/".format(wname)
+        tmp_path = "/tmp/{}".format(name)
+
+        put(archive_path, "/tmp/")
+        run("mkdir -p {}".format(releases_path))
+        run("tar -xzf {} -C {}".format(tmp_path, releases_path))
+        run("rm {}".format(tmp_path))
+        run("mv {}web_static/* {}".format(releases_path, releases_path))
+        run("rm -rf {}web_static".format(releases_path))
+        run("rm -rf /data/web_static/current")
+        run("ln -s {} /data/web_static/current".format(releases_path))
         print("New version deployed!")
         return True
     except:
-        print("New version not deployed...")
         return False
